+++ applications/tokenHUDSettings.js
@@ -17,10 +17,13 @@ export default class TokenHUDSettings extends FormApplication {
 
     async getData(options) {
         const data = super.getData(options);
+        data.tokenHUDWildcardActive = game.modules.get("token-hud-wildcard")?.active;
         const settings = game.settings.get("token-variants", "hudSettings");
         settings.enableWorldSettings = game.user && game.user.can("FILES_BROWSE") && game.user.can("TOKEN_CONFIGURE");
         if(settings.enableWorldSettings){
             settings.enableSideMenuForAll = game.settings.get("token-variants", "enableTokenHUDButtonForAll");
+            settings.displaySharedOnly = game.settings.get("token-variants", "displayOnlySharedImages");
+            settings.disableSideMenuIfTHW = game.settings.get("token-variants", "disableSideMenuIfTHW");
         }
         return mergeObject(data, settings);
     }
@@ -40,6 +43,14 @@ export default class TokenHUDSettings extends FormApplication {
         if("enableSideMenuForAll" in formData){
             game.settings.set("token-variants", "enableTokenHUDButtonForAll", formData.enableSideMenuForAll);
             delete formData.enableSideMenuForAll;
+        }
+        if("displaySharedOnly" in formData){
+            game.settings.set("token-variants", "displayOnlySharedImages", formData.displaySharedOnly);
+            delete formData.displaySharedOnly;
+        }
+        if("disableSideMenuIfTHW" in formData){
+            game.settings.set("token-variants", "disableSideMenuIfTHW", formData.disableSideMenuIfTHW);
+            delete formData.disableSideMenuIfTHW;
         }
         game.settings.set("token-variants", "hudSettings", formData);
     }